+++ core/src/commonMain/kotlin/cz/vutbr/fit/interlockSim/sim/SimpleIntegration.kt
@@ -14,7 +14,7 @@ import cz.hovorka.kdisco.Variable
 
 /**
  * Substitute for "SIMLIB Integrator"
- * Integrates dx.state into x.state using jDisco continuous simulation.
+ * Integrates dx.state into x.state using kDisco continuous simulation.
  *
  * @param x integrator output (position)
  * @param dx integrator input (velocity)