+++ core/src/jvmMain/kotlin/cz/vutbr/fit/interlockSim/util/PlatformSleep.kt
@@ -1,0 +1,9 @@
+package cz.vutbr.fit.interlockSim.util
+
+actual fun platformSleep(millis: Long) {
+	try {
+		Thread.sleep(millis)
+	} catch (_: InterruptedException) {
+		Thread.currentThread().interrupt()
+	}
+}